Added new structure of repo
